ui enhancements
