+++ frontend/app/customers/page.tsx
@@ -33,8 +33,8 @@ export default async function Page() {
       <AppSidebar variant="inset" />
       <SidebarInset>
         <SiteHeader title="Customers" />
-        <div className="flex flex-1 flex-col">
-          <div className="@container/main flex flex-1 flex-col gap-2">
+        <div className="flex flex-col overflow-hidden">
+          <div className="@container/main flex flex-col gap-2">
             <div className="flex flex-col gap-4 py-4 md:gap-6 md:py-6">
               <DataTable data={customers} />
             </div>
